// src/modules/category/category.interface.ts
// 2. Category Model:
// Fields:
// _id (Object ID): A distinctive identifier generated by MongoDB.
// name (String): A unique name of the category.

import { ObjectId } from 'mongoose'

// Create Interface for Category

export type TCategory = {
  name: string
  createdBy: {
    _id: ObjectId
    username: string
    email: string
    role: string
  }
  createdAt: Date
  updatedAt: Date
}